♻️ Refactor: Separate models and logic into distinct directories
**Directory Structure:**
- `src/models/` - All TypeScript interfaces and types
- `src/schemas/` - Zod schema definitions
- `src/tools/` - Business logic functions only

**Changes:**
- Extract TestCase, ValidationResult types to `models/test-case.ts`
- Extract ParseResult, ProcessedSteps types to `models/parse-result.ts`
- Extract ListResult, TestCaseFile types to `models/list-result.ts`
- Add common result patterns to `models/common.ts`
- Move Zod schemas to `schemas/test-case-schema.ts`
- Create centralized exports via `models/index.ts`
- Add public API exports via `src/index.ts`
- Maintain backward compatibility with re-exports
- Update all tool imports to use new model structure

**Benefits:**
- Clear separation of concerns (models vs logic)
- Centralized type definitions for better maintainability
- Improved code organization and readability
- Enhanced reusability across modules
- All 39 tests still passing with 84.65% coverage
