+++ src/index.ts
@@ -1,0 +1,12 @@
+// Public API exports for the manual tests MCP server
+
+// Export all types
+export * from './models';
+
+// Export tool functions
+export { validateTestCase } from './tools/manual-test-validate';
+export { parseTestCase } from './tools/manual-test-parse';
+export { listTestCases } from './tools/manual-test-list';
+
+// Export schemas for external validation
+export * from './schemas/test-case-schema';
