+++ src/tools/manual-test-validate.ts
@@ -1,60 +1,9 @@
 import * as yaml from 'js-yaml';
-import { z } from 'zod';
+import { TestCase, ValidationResult } from '../models';
+import { TestCaseSchema } from '../schemas/test-case-schema';
 
-// Type definitions
-export type Priority = 'high' | 'medium' | 'low';
-
-export interface TestCase {
-  meta: {
-    id: string;
-    title: string;
-    feature?: string;
-    priority: Priority;
-    tags?: string[];
-    author?: string;
-    lastUpdated?: string | Date;
-  };
-  precondition?: string[];
-  scenario: {
-    given: string[];
-    when: string[];
-    then: string[];
-  };
-  notes?: string;
-}
-
-export interface ValidationResult {
-  isValid: boolean;
-  errors: string[];
-  warnings: string[];
-  parsedData?: TestCase;
-}
-
-// Schema definitions
-const MetaSchema = z.object({
-  id: z.string().regex(/^TC-[A-Z-]+-\d+$/, 'ID format must be TC-[A-Z-]+-[NUMBER]'),
-  title: z.string().min(1, 'Title is required'),
-  feature: z.string().optional(),
-  priority: z.enum(['high', 'medium', 'low'], {
-    errorMap: () => ({ message: 'Priority must be one of: high, medium, low' })
-  }),
-  tags: z.array(z.string()).optional(),
-  author: z.string().optional(),
-  lastUpdated: z.union([z.string(), z.date()]).optional()
-});
-
-const ScenarioSchema = z.object({
-  given: z.array(z.string()).min(1, 'Given cannot be empty'),
-  when: z.array(z.string()).min(1, 'When cannot be empty'),
-  then: z.array(z.string()).min(1, 'Then cannot be empty')
-});
-
-const TestCaseSchema = z.object({
-  meta: MetaSchema,
-  precondition: z.array(z.string()).optional(),
-  scenario: ScenarioSchema,
-  notes: z.string().optional()
-});
+// Re-export types for backward compatibility
+export type { TestCase, ValidationResult, Priority } from '../models';
 
 /**
  * Parse YAML content safely
